requirements.txt
requests
faiss-cpu
sentence-transformers
python-dotenv

